@@ -5,7 +5,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.3'
+        classpath 'com.android.tools.build:gradle:3.3.1'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -25,9 +25,9 @@ allprojects {
 
 
 ext {
-  buildToolsVersion = "26.0.3"
+  buildToolsVersion = "28.0.3"
   minSdkVersion = 21
-  compileSdkVersion = 26
-  targetSdkVersion = 26
-  supportLibVersion = "26.1.0"
+  compileSdkVersion = 28
+  targetSdkVersion = 28
+  supportLibVersion = "1.0.2"
 }
